impl small rtti and ecs basis

+++ src/rtti/rtti.cpp
@@ -4,7 +4,7 @@
 
 namespace arena
 {
-	uint64 RTTIData::s_IDGenerator = 0u;
+	uint32 RTTIData::s_IDGenerator = 0u;
 
 	RTTIData::RTTIData(const String tname) : m_typename(tname),
 											 m_typeID(s_IDGenerator++) {
